Add UI for listMeeting
There are significant changes in CommandResults to take note of: 2 new
attributes are defined in order to denote whether the command was client
or meeting specific. These are not in a single attribute as there can be
commands that are not client or meeting specific and we don't want to
change the view unneccessarily (to switch to meeting or client specific).

The change will require all existing commands to use the new constructor
in order for UI to work properly for the command.

Unfortunately, because of the intrinsic behaviour of method overloading,
the constructor for CommandResult is now long with 2 additional
parameters. One solution could be to use factory methods to create
CommandResults. Another could be to use a more awkward parameter
positions (`public CommandResult(boolean isClient, boolean isMeeting,
String feedbackToUser)`).

The logic for switching of panels can be found in MainWindow.java. To
summarise, we simply clear the placeholder and add the panel we need to
the placeholder each time we swap.